+++ README.md
@@ -1,1 +1,1 @@
-# Hangman
+# Hangman game implement to HTML exercise
